GetUser method

+++ store/datastore/users.go
@@ -9,7 +9,7 @@ func (db *Datastore) GetUserList() ([]*model.User, error) {
 	return users, err
 }
 
-func (db *Datastore) GetUser(id int64) (*model.User, error) {
+func (db *Datastore) GetUser(id uint) (*model.User, error) {
 	var user = &model.User{}
 	var err = db.First(&user, id).Error
