+++ store/store.go
@@ -8,7 +8,7 @@ import (
 
 type Store interface {
 	GetUserList() ([]*model.User, error)
-	GetUser(id int64) (*model.User, error)
+	GetUser(id uint) (*model.User, error)
 }
 
 // GetUserList gets a list of all users in the system.
@@ -16,6 +16,7 @@ func GetUserList(c context.Context) ([]*model.User, error) {
 	return FromContext(c).GetUserList()
 }
 
-func GetUser(c context.Context, id int64) (*model.User, error) {
+// GetUser gets the user object corresponding to the given id.
+func GetUser(c context.Context, id uint) (*model.User, error) {
 	return FromContext(c).GetUser(id)
 }
